Support Vector Regression  - Truth or Bluff
Trying to find the better algorithm to use in Truth or Bluff case, we applied the SVR model using Gausian Radial Basis Function (RBF) as kernel.
The result wasn't good, once the model couldn't fit high values, values which could be an outlier. It's possible yet applying Gradient Method to tunning it, but will be tryed another algorithms that could fit better than SVR model for this case.