Decision tree regression - Study Case
This study case was used to learn about the decision tree algorithm and your methods.
I applied it to discover about a candidate salary, for a position job. Analysing  plethora of positions and salaries we could identify if the candidate tell us a truth or bluff about his previous salary.