Hands Washing Case Study
In this project we analyse the time when was discovered the influence of won't washing hands on many cases of deaths. 
We analysing the proportion of deaths on births before and after doctors be obligatorie to wash their hands for make surgerys.
We use a bootstrap method to analyse the reduction of deaths , and analyse the  confidence interval from the difference of mean between before and after the Hands Washing event.